Fix warnings from merge

+++ src/Chirp.Web/Pages/AboutMe.cshtml.cs
@@ -138,6 +138,7 @@ public class AboutMeModel : PageModel
             return RedirectToPage();
         }
     }
+
     public async Task<IActionResult> OnPostSetDarkMode()
     {
         var isDarkMode = !await _authorRepository.IsDarkMode(User.Identity?.Name!);
@@ -150,13 +151,15 @@ public class AboutMeModel : PageModel
     {
         if (scale == 15)
         {
-            scale = (float)1.5;
+            scale = 1.5F;
         }
+
         if (scale < 1 || scale > 2)
         {
             Console.WriteLine($"Invalid scale: {scale}");
             return RedirectToPage();
         }
+
         Console.WriteLine(scale);
         await _authorRepository.SetFontSizeScale(User.Identity?.Name!, scale);
         return RedirectToPage();
